[DF-12085] fixed error on launcher startup
Former-commit-id: 66f51328e5039c6575c754d03704a409206f6577

+++ Tools/Launcher/Classes/main.cpp
@@ -26,18 +26,21 @@ int main(int argc, char* argv[])
     QDir currentDir(fi.absoluteDir());
     QString platformsPath = "platforms";
     QString windowsDllPath = "qwindows.dll";
-    //remove "platforms" with permission hack
-    if (currentDir.exists(platformsPath))
+    if (currentDir.exists(windowsDllPath))
     {
-        currentDir.cd(platformsPath);
-        currentDir.removeRecursively();
-        currentDir.cdUp();
-    }
-    //try copy even if folder already exists
-    //if (!currentDir.exists(platformsDir))
-    {
-        currentDir.mkpath(platformsPath);
-        QFile::copy(windowsDllPath, platformsPath + "/" + windowsDllPath);
+        //remove "platforms" with permission hack
+        if (currentDir.exists(platformsPath))
+        {
+            currentDir.cd(platformsPath);
+            currentDir.removeRecursively();
+            currentDir.cdUp();
+        }
+        //try copy even if folder already exists
+        //if (!currentDir.exists(platformsDir))
+        {
+            currentDir.mkpath(platformsPath);
+            QFile::copy(windowsDllPath, platformsPath + "/" + windowsDllPath);
+        }
     }
 #endif //windows
     QApplication a(argc, argv);
